Cloudtop updated to use UID 1000 for VNC user
Updated cloudtop image versions for related apps.  The VNC user’s UID was changed from using the default (currently 999) UID to using one that is set in the Dockerfile (USER_ID=1000).  This way the UID is the same as the one used in other images and should have less problems with shared filesystems on PVCs.